+++ epicure-admin/src/app/components/restaurants/restaurants.component.html
@@ -1,1 +1,5 @@
-<p>restaurants works!</p>
+<div class="container-lg">
+    <button type="button" class="btn btn-primary">
+      New Restaurant
+    </button>
+
