+++ epicure-admin/src/app/components/restaurants/restaurants.component.scss
@@ -1,0 +1,7 @@
+.table-responsive {
+    width: 100%;
+  
+    .restaurant-row {
+      cursor: pointer;
+    }
+  }
