+++ epicure-admin/src/app/components/restaurants/restaurants.component.ts
@@ -1,6 +1,4 @@
-import { Component, ViewChild } from '@angular/core';
-import { NgModule } from '@angular/core';
-
+import { Component } from '@angular/core';
 import { IRestaurant } from '../../interface/restaurant.interface';
 import { RestaurantsService } from '../../service/restaurant.service';
 
@@ -11,7 +9,21 @@ import { RestaurantsService } from '../../service/restaurant.service';
   styleUrl: './restaurants.component.scss'
 })
 export class RestaurantsComponent {
-  private restaurants: IRestaurant[] = [];
-  
-  
+  restaurants!: IRestaurant[];
+  constructor(private restaurantsService: RestaurantsService) { }
+
+  ngOnInit(): void {
+    this.restaurantsService.restaurantsUpdateEvent.subscribe((restaurants: IRestaurant[]) => {
+      this.restaurants = restaurants;
+    })
+  }
+
+  ngAfterContentInit() {
+    this.setRestaurants();
+  }
+
+  setRestaurants() {
+    this.restaurants = this.restaurantsService.restaurants;
+  }
+
 }
