+++ epicure-admin/src/app/service/restaurant.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@angular/core';
+import { EventEmitter, Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { IRestaurant } from '../interface/restaurant.interface';
@@ -10,27 +10,35 @@ import { IRestaurant } from '../interface/restaurant.interface';
 export class RestaurantsService {
     private apiUrl = 'http://localhost:3000/api' + '/restaurant';
    
-    constructor(private http: HttpClient) { }
+    restaurantsUpdateEvent = new EventEmitter<IRestaurant[]>();
 
-    getAllRestaurants(): Observable<IRestaurant[]> {
-        return this.http.get<IRestaurant[]>(this.apiUrl);
+    restaurants: IRestaurant[] = [];
+    constructor(private http: HttpClient) {
+        this.fetchRestaurants();
     }
 
-    // getRestaurantsAdmin(): Observable<IRestaurant[]> {
-    //     return this.http.get<IRestaurant[]>(this.apiAdminUrl);
+    fetchRestaurants() {
+        //need to check why using API_URL.chefs not working
+        this.http.get(this.apiUrl).subscribe((restaurants) => {
+            console.log("Restaurants", restaurants)
+            //this.restaurants = restaurants;
+            // this.restaurantsUpdateEvent.emit(restaurants);
+        })
+    }
+
+    // updateRestaurant(restaurant: IRestaurant) {
+    //     //need to check why using API_URL.chefs not working
+    //     this.http.put(`http://127.0.0.1:3000/restaurants/${restaurant._id}`, restaurant).subscribe((responseRestaurant: IRestaurant) => {
+    //         console.log("response put restaurant", responseRestaurant);
+    //         this.fetchRestaurants();
+    //     })
     // }
 
-    updateRestaurantById(restaurantData: IRestaurant): Observable<IRestaurant> {
-        const url = `${this.apiUrl}/${restaurantData._id}`;
-        return this.http.put<IRestaurant>(url, restaurantData);
-    }
-
-    deleteRestaurantById(restaurantData: IRestaurant): Observable<IRestaurant> {
-        const url = `${this.apiUrl}/${restaurantData._id}`;
-        return this.http.delete<IRestaurant>(url, restaurantData);
-    }
-
-    addNewRestaurant(restaurantData: IRestaurant): Observable<IRestaurant> {
-        return this.http.post<IRestaurant>(this.apiUrl, restaurantData);
-    }
+    // postRestaurant(restaurant: IRestaurant) {
+    //     //need to check why using API_URL.chefs not working
+    //     this.http.post('http://127.0.0.1:3000/restaurants', restaurant).subscribe((responseRestaurant: IRestaurant) => {
+    //         console.log("response post restaurant", responseRestaurant);
+    //         this.fetchRestaurants();
+    //     })
+    // }
 }
